refactor: Move function parsers to their own module
